set mesh verts to those of the initial frame
for group frame models the mesh data was just set with the initial frame
of group frame 0, rather than using initial_pose_num.  for some reason
this conflicts if the shape keys are from a different group frame.  this
can be seen in flame2 static models with pose num 1, where the width of
the flame fluctuates wildly in comparison to how it appears in
Quakespasm.  with the fix the flame is more stable and resembles
Quakespasm more closely